+ Shoot Behaviour (WIP) * UI recolor death menu

+++ Assets/Scripts/ScriptableUpdaters/Templates/PlayerData.cs
@@ -8,6 +8,6 @@ public class PlayerData : InstantiableScriptable
     public float speed = 10;
 	public float gravity = 12;
     public float jumpSpeed = 0;
-    public float ammoAmount = 0;
+    public int ammoAmount = 0;
     public bool isInvincible = false;
 }
